fix: netlify build fix
